ETCM-167: Moved tests to separate files.
